% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_pcd.R
\name{collapse_pcd}
\alias{collapse_pcd}
\title{Collapse Senamhi stations with common variables}
\usage{
collapse_pcd(datain)
}
\arguments{
\item{datain}{a list of individual station tables acquired through the \code{senamhiR} function.}
}
\value{
a list of collapsed stations
}
\description{
Collapse Senamhi stations with common variables
}
