% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadAction.R
\name{downloadAction}
\alias{downloadAction}
\title{Curl helper}
\usage{
downloadAction(url, filename, writeMode = NULL)
}
\arguments{
\item{url}{character; address to be downloaded.}

\item{filename}{character; name to save the downloaded file under.}

\item{writeMode}{character; if set to 'overwrite' the script will overwrite file if it exists.}
}
\value{
None
}
\description{
A helper function to execute download actions using curl.
}
\author{
Conor I. Anderson
}

