% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{downloadData}
\alias{downloadData}
\title{Download from the Peruvian National Hydrological and Meterological Service}
\usage{
downloadData(station, year, month = 1:12, writeMode = "z")
}
\arguments{
\item{station}{character; the station id number to process.}

\item{year}{numerical; a vector of years to process.}

\item{month}{numerical; a vector of months to process. Defaults to 1:12.}

\item{writeMode}{character; if set to 'overwrite', the script will overwrite downloaded files if they exist.}
}
\value{
None
}
\description{
Download Peruvian historical climate data from the Senamhi web portal.
}
\examples{
\dontrun{downloadData("000401", 1971:2000)}
}
\author{
Conor I. Anderson
}

