% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{export_data}
\alias{export_data}
\title{Compile data from the Peruvian National Hydrological and Meterological Service}
\usage{
export_data(station, year, month = 1:12, write_mode = "z", trim = TRUE)
}
\arguments{
\item{station}{character; the station id number to process.}

\item{year}{numerical; a vector of years to process.}

\item{month}{numerical; a vector of months to process. Defaults to 1:12.}

\item{write_mode}{character; if set to 'append', the script will append the data to an exisiting file; if set to 'overwrite', it will overwrite an existing file. If not set, it will not overwrite.}

\item{trim}{logical; if set to TRUE, the script will trim missing data from the start and end of the data set. Note only completely missing years will be trimmed.}
}
\value{
None
}
\description{
Compile a CSV file of Peruvian historical climate data from the Senamhi web portal.
}
\examples{
\dontrun{export_data('000401', 2000:2005, 1:12, trim = TRUE)}
}
\author{
Conor I. Anderson
}

