% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_local_catalogue.R
\name{generate_local_catalogue}
\alias{generate_local_catalogue}
\title{Generate Local Catalogue}
\usage{
generate_local_catalogue(station, localcatalogue)
}
\arguments{
\item{station}{character; optional character string to specify one or more stations.}

\item{localcatalogue}{character; optional character string to specify catalogue object to update. by region.}
}
\value{
A data frame containing the details of matching stations.
}
\description{
Generate or update a catalogue of locally downloaded stations.
}
\examples{
# Update catalogue information for 'Tarapoto'.
\dontrun{generate_local_catalogue("000401")}

}
\author{
Conor I. Anderson
}

