% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCatalogue.R
\name{generateCatalogue}
\alias{generateCatalogue}
\title{A function to determine the full catalogue of available Peruvian National Hydrological and Meterological Service stations}
\usage{
generateCatalogue()
}
\value{
catalogue.rda
}
\description{
Generate a .rda file containing a list of all of the stations operated by Senamhi. You should not need to execute this function, as the data is already included in the package.
}
\examples{
generateCatalogue()
}
\author{
Conor I. Anderson
}

