% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessConfig.R
\name{guessConfig}
\alias{guessConfig}
\title{Guess station characteristics}
\usage{
guessConfig(station, writeMode = "z")
}
\arguments{
\item{station}{character; the number of the station id number to process.}

\item{writeMode}{character; if set to 'overwrite', the script will overwrite downloaded files if they exist.}
}
\value{
vector
}
\description{
Attempt to guess station characteristics.
}
\examples{
\dontrun{guessConfig("000401")}
}
\author{
Conor I. Anderson
}

