% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessPeriod.R
\name{guessPeriod}
\alias{guessPeriod}
\title{Query available data from the Peruvian National Hydrological and Meterological Service}
\usage{
guessPeriod(station, automatic = TRUE, writeMode = NULL)
}
\arguments{
\item{station}{character; the number of the station id number to process.}

\item{automatic}{logical; if set to true (default), the script will attempt to guess the startYear and endYear values.}

\item{writeMode}{character; if set to 'overwrite', the script will overwrite downloaded files if they exist.}
}
\value{
data.frame
}
\description{
Query the available data for a given station from the Senamhi web portal.
}
\examples{
\dontrun{guessPeriod("000401")}
}
\author{
Conor I. Anderson
}

