% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_stations.R
\name{map_stations}
\alias{map_stations}
\title{Map Senamhi stations on an interactive map}
\usage{
map_stations(station, type = "osm")
}
\arguments{
\item{station}{character; one or more station id numbers to show on the map.}

\item{type}{character; either "osm" for OpenStreetMap tiles, or "sentinel" for cloudless satellite by EOX IT Services GmbH (\url{https://s2maps.eu}).}
}
\description{
Show the stations of interest on an interactive map using Leaflet. Zoom levels are guessed based on an RStudio plot window.
}
\examples{
# Map a single station
map_stations(401)
# Make a map from station search results.
map_stations(station_search(region = "SAN MARTIN", baseline = 1981:2010))
}
\author{
Conor I. Anderson
}
