% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qc}
\alias{qc}
\title{Basic data quality control}
\usage{
qc(dat)
}
\arguments{
\item{dat}{a \code{tbl_df} generated form the \code{senamhiR} package}
}
\value{
a \code{tbl_df}
}
\description{
A helper function to perform minimal quality control on the data. 
For now, this script only performs action on the three main temperature variables.
}
\author{
Conor I. Anderson
}
