% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senamhi.R
\name{senamhi}
\alias{senamhi}
\title{A wrapper function to download and compile data from the Peruvian National Hydrological and Meterological Service}
\usage{
senamhi(tasks, station, automatic = TRUE, dataAvail = TRUE, type = "z",
  MorH = "z", startYear, endYear, startMonth = 1, endMonth = 12,
  overwrite = FALSE, append = FALSE, custom = FALSE)
}
\arguments{
\item{tasks}{numerical; define which tasks to perform: 1) Download Data, 2) Compile CSV of Downloaded Data, 3) Both.}

\item{station}{character; the number of the station id number to process. Can also be a vector of station ids.}

\item{automatic}{logical; if set to true (default), the script will attempt to guess the type and MorH values as well as startYear and endYear}

\item{dataAvail}{logical; if set to true (default), the script will either automatically attempt to choose start and end datas, or return a(n outdated) table of the data available for your station.}

\item{type}{character; defines if the station is (CON)ventional, DAV, (SUT)ron, or (SIA)p. Must be "CON", "DAV", "SUT" or "SIA".}

\item{MorH}{character; defines if the station is (M)eterological (2) or (H)ydrological. Must be "M", "M2" or "H".}

\item{startYear}{numerical; the first year to process.}

\item{endYear}{numerical; the last year to process.}

\item{startMonth}{numerical; the first month to process. Defaults to 1.}

\item{endMonth}{numerical; the last month to process. Defaults to 12.}

\item{overwrite}{logical; if true, the script will overwrite downloaded files if they exist.}

\item{append}{logical; if true, the script will append the data to an exisiting file, otherwise it will overwrite.}

\item{custom}{logical; if true, the script will provide the opportunity to manually enter column headers.}
}
\value{
None
}
\description{
Download and/or compile Peruvian historical climate data from the Senamhi web portal.
}
\examples{
senamhi()
senamhi(3, 000401, type = "CON", MorH = "M", 1971, 2000)
}
\author{
Conor I. Anderson
}

