% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senamhiCatalogue.R
\name{senamhiCatalogue}
\alias{senamhiCatalogue}
\title{A function to determine the full catalogue of available Peruvian National Hydrological and Meterological Service stations}
\usage{
senamhiCatalogue()
}
\value{
catalogue.rda
}
\description{
Generate a .rda file containing a list of all of the stations operated by Senamhi.
}
\examples{
senamhiCatalogue()
}
\author{
Conor I. Anderson
}

