% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senamhiDownload.R
\name{senamhiDownload}
\alias{senamhiDownload}
\title{Download from the Peruvian National Hydrological and Meterological Service}
\usage{
senamhiDownload(station, type = "z", MorH = "z", startYear, endYear,
  startMonth = 1, endMonth = 12)
}
\arguments{
\item{station}{numerical; the number of the station id number to process.}

\item{type}{character; defines if the station is (CON)ventional, (SUT)ron, or (SIA)p. Must be "CON", "SUT" or "SIA".}

\item{MorH}{character; defines if the station is (M)eterological or (H)ydrological. Must be "M" or "H".}

\item{startYear}{numerical; the first year to process.}

\item{endYear}{numerical; the last year to process.}

\item{startMonth}{numerical; the first month to process. Defaults to 1.}

\item{endMonth}{numerical; the last month to process. Defaults to 12.}
}
\value{
None
}
\description{
Download Peruvian historical climate data from the Senamhi web portal.
}
\examples{
senamhiDownload()
senamhiDownload(000401, type = "CON", MorH = "M", 1971, 2000)
}
\author{
Conor I. Anderson
}

