% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senamhiGuess.R
\name{senamhiGuess}
\alias{senamhiGuess}
\title{Guess station characteristics}
\usage{
senamhiGuess(station)
}
\arguments{
\item{station}{numerical; the number of the station id number to process.}
}
\value{
vector
}
\description{
Attempt to guess station characteristics.
}
\examples{
senamhiGuess()
senamhiGuess("000401")
}
\author{
Conor I. Anderson
}

