% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortFiles.R
\name{sortFiles}
\alias{sortFiles}
\title{Sort downloaded and compiled data}
\usage{
sortFiles(station)
}
\arguments{
\item{station}{character; the number of the station id number to process.}
}
\value{
none
}
\description{
Sort downloaded html and compiled csv files into folders sorted by region.
}
\examples{
\dontrun{sortFiles("000401")}
}
\author{
Conor I. Anderson
}

