% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_explorer.R
\name{station_explorer}
\alias{station_explorer}
\title{A Shiny interface to Senamhi weather and river stations}
\usage{
station_explorer()
}
\value{
none
}
\description{
A function to launch a shiny web app to explore Senamhi stations.
}
\author{
Conor I. Anderson
}

