% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_search.R
\name{station_search}
\alias{station_search}
\title{Find Senamhi stations matching various criteria}
\usage{
station_search(name = NULL, ignore.case = TRUE, glob = FALSE,
  region = NULL, period = NULL, config = NULL, target = NULL,
  dist = 0:100, sort = TRUE, ...)
}
\arguments{
\item{name}{character; optional character vector to filter results by station name.}

\item{ignore.case}{logical; by default the search for station names is not case-sensitive.}

\item{glob}{logical; whether to allow regular expressions in the \code{name}. See \code{\link{glob2rx}}.}

\item{region}{character; optional character string to filter results by region.}

\item{period}{numeric; optional, either a range of years or the total number of years of data that must be available.}

\item{config}{character; the configuration of the station ((m)eteorological or (h)ydrological)}

\item{target}{numeric; optional station ID of a target station, or a vector of length 2 containing latitude and longitude (in that order).}

\item{dist}{numeric; vector with a range of distance from the target in km. Only used if a target is specified. (default is 0:100)}

\item{sort}{Boolean; if TRUE (default), will sort the resultant table by distance from `target`. Only used if a target is specified.}

\item{...}{Additional arguments passed to \code{\link{grep}}.}
}
\value{
A data frame containing the details of matching stations.
}
\description{
Search for Senamhi stations by name, region, available data, and/or distance to a target.
}
\examples{
# Find all stations containing 'Tarapoto' in their name.
station_search('Tarapoto')

# Find all stations starting with "San"
station_search(name = "San*", glob = TRUE)

# Find stations with data available from 1971 to 2000.
station_search(period = 1971:2000)

# Find all stations between 0 and 100 km from Station '000401'
station_search(target = '000401', dist = 0:100)

}
\author{
Conor I. Anderson
}
