% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationExporerGUI.R
\name{stationExplorerGUI}
\alias{stationExplorerGUI}
\title{A Shiny interface to Senamhi weather and river stations}
\usage{
stationExplorerGUI()
}
\value{
none
}
\description{
A function to launch a shiny web app to explore Senamhi stations.
}

