% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationSearch.R
\name{stationSearch}
\alias{stationSearch}
\title{Find Senamhi stations matching various criteria}
\usage{
stationSearch(name = NULL, region = NULL, baseline = NULL,
  ignore.case = TRUE, target = NULL, mindist = 0, maxdist = 100,
  sort = TRUE, ...)
}
\arguments{
\item{name}{character; optional character string to filter results by station name.}

\item{region}{character; optional character string to filter results by region.}

\item{baseline}{vector; optional vector with a start and end year for a desired baseline.}

\item{ignore.case}{logical; by default the search for station names is not case-sensitive.}

\item{target}{numeric; optional station ID of a target station, or a vector of length 2 containing latitude and longitude (in that order).}

\item{mindist}{numeric; minimum distance from the target in km. Only used if a target is specified. (defaults to 0)}

\item{maxdist}{numeric; maximum distance from the target in km. Only used if a target is specified. (defaults to 100)}

\item{sort}{Boolean; if TRUE (default), will sort the resultant table by distance from `target`. Only used if a target is specified.}

\item{...}{Additional arguments passed to \code{\link{grep}}.}
}
\value{
A data frame containing the details of matching stations.
}
\description{
Search for Senamhi stations by name, region, available data, and/or distance to a target.
}
\examples{
# Find all stations containing "Tarapoto" in their name.
stationSearch("Tarapoto")

# Find stations with data available from 1971 to 2000.
stationSearch(baseline = c(1971, 2000))

# Find all stations between 0 and 100 km from Station "000401"
stationSearch(target = "000401", mindist = 0, maxdist = 100)

}
\author{
Conor I. Anderson
}

