% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_data}
\alias{write_data}
\title{Compile data from the Peruvian National Hydrological and Meterological Service}
\usage{
write_data(station, year, write_mode = "z", trim = TRUE, clean = FALSE)
}
\arguments{
\item{station}{character; the station id number to process.}

\item{year}{numerical; a vector of years to process.}

\item{write_mode}{character; if set to 'append', the script will append the data to an exisiting file; if set to 'overwrite', it will overwrite an existing file. If not set, it will not overwrite.}

\item{trim}{logical; if set to TRUE, the script will trim missing data from the start and end of the data set. Note only completely missing years will be trimmed.}

\item{clean}{logical; if set to TRUE, the script will delete all of the downloaded HTML files.}
}
\value{
None
}
\description{
Compile a CSV file of Peruvian historical climate data from the Senamhi web portal.
}
\examples{
\dontrun{write_data('000401', 2000:2005, trim = TRUE)}
}
\author{
Conor I. Anderson
}

