% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catalogue}
\alias{catalogue}
\title{Catalogue of Senamhi stations}
\format{An object of class \code{data.frame} with 1789 rows and 12 columns.}
\source{
\url{http://www.senamhi.gob.pe/}
}
\usage{
catalogue
}
\description{
A \code{data.frame} containing the identifiers, names, and other characteristics of climate and hydro stations in Peru.
}
\keyword{datasets}

