% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Access data from the Peruvian National Hydrological and Meterological Service via API}
\usage{
download_data(station, year = NULL)
}
\arguments{
\item{station}{character; the station id number to process.}

\item{year}{numeric; an ordered vector of years to retrieve.}
}
\value{
tbl_df
}
\description{
Download archived Peruvian historical climate data from the Senamhi via an independent API.
}
\examples{
\dontrun{download_data('000401')}
}
\author{
Conor I. Anderson
}
