% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_stations.R
\name{map_stations}
\alias{map_stations}
\title{Map Senamhi stations on an interactive map}
\usage{
map_stations(station, zoom)
}
\arguments{
\item{station}{character; one or more station id numbers to show on the map.}

\item{zoom}{numeric; the level to zoom the map to.}
}
\description{
Show the stations of interest on an interactive map using Leaflet. Zoom levels are guessed based on an RStudio plot window.
}
\examples{
# Make a map of all the stations.
\dontrun{map_stations(catalogue$StationID), zoom = 4}
# Make a map of all stations in Cusco.
\dontrun{map_stations(catalogue$StationID[catalogue$Region == "CUSCO"])}
}
\author{
Conor I. Anderson
}

