% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read data from the Peruvian National Hydrological and Meterological Service}
\usage{
read_data(station, path = "default")
}
\arguments{
\item{station}{character; the station id number to process.}

\item{path}{character; the path to the file. By default, the script expects the csv files to be sorted by region.}
}
\value{
A tibble (tbl_df) of the relative information.
}
\description{
Read a CSV file of Peruvian historical climate data from the Senamhi web portal.
}
\examples{
\dontrun{read_data('000401')}
}
\author{
Conor I. Anderson
}
