% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senamhiGetPeriod.R
\name{senamhiGetPeriod}
\alias{senamhiGetPeriod}
\title{Query available data from the Peruvian National Hydrological and Meterological Service}
\usage{
senamhiGetPeriod(station)
}
\arguments{
\item{station}{numerical; the number of the station id number to process.}
}
\value{
data.frame
}
\description{
Query the available data for a given station from the Senamhi web portal.
}
\examples{
senamhiGetPeriod()
senamhiGetPeriod(000401)
}
\author{
Conor I. Anderson
}

