% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_explorer.R
\name{station_explorer}
\alias{station_explorer}
\title{A Shiny interface to Senamhi weather and river stations}
\usage{
station_explorer(local = FALSE)
}
\arguments{
\item{local}{logical; if set to `TRUE`, we will show only the data that is available locally.}
}
\value{
none
}
\description{
A function to launch a shiny web app to explore Senamhi stations.
}
\examples{
\dontrun{station_explorer()}
}
\author{
Conor I. Anderson
}

