% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCSV.R
\name{writeCSV}
\alias{writeCSV}
\title{Compile data from the Peruvian National Hydrological and Meterological Service}
\usage{
writeCSV(station, year, month = 1:12, writeMode = "z")
}
\arguments{
\item{station}{character; the station id number to process.}

\item{year}{numerical; a vector of years to process.}

\item{month}{numerical; a vector of months to process. Defaults to 1:12.}

\item{writeMode}{character; if set to 'append', the script will append the data to an exisiting file; if set to 'overwrite', it will overwrite an existing file. If not set, it will not overwrite.}
}
\value{
None
}
\description{
Compile a CSV file of Peruvian historical climate data from the Senamhi web portal.
}
\examples{
\dontrun{writeCSV("000401", 2000:2005, 1:12)}
}
\author{
Conor I. Anderson
}

