% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCSV.R
\name{writeCSV}
\alias{writeCSV}
\title{Compile data from the Peruvian National Hydrological and Meterological Service}
\usage{
writeCSV(station, type = "z", config = "z", startYear, endYear,
  startMonth = 1, endMonth = 12, overwrite = FALSE, append = FALSE,
  custom = FALSE)
}
\arguments{
\item{station}{character; the number of the station id number to process.}

\item{type}{character; defines if the station is (CON)ventional, DAV, (SUT)ron, or (SIA)p. Must be "CON", "DAV", "SUT" or "SIA".}

\item{config}{character; defines if the station is (M)eterological (2) or (H)ydrological. Must be "M", "M2" or "H".}

\item{startYear}{numerical; the first year to process.}

\item{endYear}{numerical; the last year to process.}

\item{startMonth}{numerical; the first month to process. Defaults to 1.}

\item{endMonth}{numerical; the last month to process. Defaults to 12.}

\item{overwrite}{logical; if true, the script will overwrite the compiled csv file if it exists.}

\item{append}{logical; if true, the script will append the data to an exisiting file, otherwise it will overwrite.}

\item{custom}{logical; if true, the script will provide the opportunity to manually enter column headers.}
}
\value{
None
}
\description{
Compile as CSV file of Peruvian historical climate data from the Senamhi web portal.
}
\examples{
writeCSV()
writeCSV("000401", type = "CON", config = "M", 1971, 2000, 1, 12)
}
\author{
Conor I. Anderson
}

